add duckdb-httpfs python package
- refactor `setup.py`
- add `httpfs_package/setup.py`
- speed up compilation for linux/darwin by using multiprocess
- hack `sdist` to reduce copy/paste amounts...
- add a github action for python3 wheels
- enforce PEP8 where possible
