+++ .github/workflows/main.yml
@@ -924,6 +924,47 @@ jobs:
         mkdir duckdb_tarball && tar xvf dist/duckdb-*.tar.gz --strip-components=1 -C duckdb_tarball
         cibuildwheel --output-dir wheelhouse duckdb_tarball
 
+  linux-python3-httpfs:
+    name: Python 3 Linux with HTTPFS
+    runs-on: ubuntu-20.04
+    needs: linux-debug
+
+    env:
+      CIBW_BUILD: 'cp36-* cp37-* cp38-* cp39-*'
+      CIBW_BEFORE_BUILD: 'yum install -y openssl-devel && pip install --prefer-binary "pandas>=0.24" "pytest>=4.3"'
+      CIBW_TEST_REQUIRES: 'pytest'
+      CIBW_BEFORE_TEST: 'yum install -y openssl && pip install --prefer-binary "pandas>=0.24" && (pip install --prefer-binary "pyarrow>=4.0.0" || true)'
+      CIBW_TEST_COMMAND: 'python -m pytest {project}/tests'
+      SETUPTOOLS_SCM_NO_LOCAL: 'yes'
+      TWINE_USERNAME: 'hfmuehleisen'
+
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - uses: actions/setup-python@v2
+        with:
+          python-version: '3.7'
+
+      - name: Install
+        run: pip install cibuildwheel twine
+
+      - name: Build
+        run: |
+          cd tools/pythonpkg/httpfs_package
+          python setup.py sdist
+          cd ..
+          mkdir duckdb_httpfs_tarball && tar xvf dist/duckdb-*.tar.gz --strip-components=1 -C duckdb_httpfs_tarball
+          cp httpfs_package/setup.py duckdb_httpfs_tarball/
+          cibuildwheel --output-dir wheelhouse duckdb_httpfs_tarball
+
+      - name: Deploy
+        run: |
+          python scripts/asset-upload-gha.py duckdb_python_src.tar.gz=tools/pythonpkg/dist/duckdb-*.tar.gz
+          if [[ "$GITHUB_REF" =~ ^(refs/heads/master|refs/tags/v.+)$ ]] ; then
+            twine upload --non-interactive --disable-progress-bar --skip-existing tools/pythonpkg/wheelhouse/*.whl tools/pythonpkg/dist/duckdb-*.tar.gz
+          fi
 
   osx-python3:
     name: Python 3 OSX
